Consolidate nth & average cone ray computation to one class
Merges AveragedConeRaySegmentComputation and
NthPercentileConeRaySegmnetComputation into
StatisticalConeRaySegmentComputation.
