--- Runtime/Components/ConeRayAnglesEstimator/AveragedConeRaySegmentComputation.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 28c2257af5eafe74dbae03b6bf3564e6
