+++ Runtime/Components/ConeRayAnglesEstimator/StatisticalConeRaySegmentComputation.cs
@@ -3,19 +3,43 @@ using System.Collections.Generic;
 using System.Linq;
 using ubco.ovilab.HPUI.Interaction;
 using UnityEngine;
-using UnityEngine.Assertions;
 
 namespace ubco.ovilab.HPUI
 {
     /// <summary>
-    /// Computes the cone ray angles by averaging accords frames.
+    /// Computes the cone ray angles for a segment using statistical estimates based on frame data.
+    /// This class takes interaction records and, for each qualified ray, calculates an estimated distance
+    /// using either an average or a specified percentile, optionally scaled by a multiplier.
+    /// Rays must be present in at least a minimum threshold of frames to qualify.
     /// </summary>
     [Serializable]
-    public class AveragedConeRaySegmentComputation : IConeRaySegmentComputation
+    public class StatisticalConeRaySegmentComputation : IConeRaySegmentComputation
     {
+        public enum Estimate {
+            /// <summary>
+            /// For each ray seen during interactions, use the average
+            /// interaction distance from all the data collected.
+            /// </summary>
+            Average,
+
+            /// <summary>
+            /// For each ray seen during interactions, get the nth
+            /// percentile distance from all the data collected.
+            /// </summary>
+            Percentile
+        }
+
+        [SerializeField]
+        [Tooltip("The statistical estimate to use.")]
+        private Estimate estimate;
+
         [SerializeField, Range(0.01f, 1f)]
         [Tooltip("The percentage of frames in the gesture that a ray should have been used to qualify for the final cone")]
         private float minRayInteractionsThreshold = 0.2f;
+
+        [SerializeField, Range(0f, 1f)]
+        [Tooltip("When Percentile is used, get the nth percentile distance at which interactions occured for each ray")]
+        public float percentile = 0.6f;
 
         [SerializeField, Range(1f, 2f)]
         [Tooltip("Multiply each ray by a fixed multiplier. Useful for when the rays produced are frequently losing contact during gestures.")]
@@ -25,6 +49,8 @@ namespace ubco.ovilab.HPUI
         {
             Dictionary<(float, float), float> averageRayDistance = new();
             // For each interaction, get the frame with the shortest distance
+            bool atLeastOneRayAnalyzed = false;
+
             foreach (ConeRayComputationDataRecord interactionRecord in interactionRecords)
             {
                 if (interactionRecord.segment == segment)
@@ -51,15 +77,21 @@ namespace ubco.ovilab.HPUI
                     int frameCountForMinRayInteractionsThreshold = (int)(minRayInteractionsThreshold * interactionRecord.records.Count);
                     foreach (var ray in rayDistances)
                     {
+                        atLeastOneRayAnalyzed = true;
                         if (ray.Value.Count > frameCountForMinRayInteractionsThreshold)
                         {
-                            averageRayDistance[(ray.Key.Item1, ray.Key.Item2)] = ray.Value.Average() * multiplier;
+                            averageRayDistance[(ray.Key.Item1, ray.Key.Item2)] = estimate switch
+                            {
+                                Estimate.Average => ray.Value.Average() * multiplier,
+                                Estimate.Percentile => ray.Value.Percentile(percentile) * multiplier,
+                                _ => throw new NotImplementedException("Unknown value")
+                            };
                         }
                     }
                 }
             }
 
-            if (averageRayDistance.Count() == 0)
+            if (atLeastOneRayAnalyzed && averageRayDistance.Count() == 0)
             {
                 Debug.LogWarning($"Data collection has gone wrong for Phalange {segment.ToString()}, no rays have been utilized enough for ray interaction threshold of {minRayInteractionsThreshold}");
             }
@@ -72,6 +104,30 @@ namespace ubco.ovilab.HPUI
             }
 
             return coneAnglesForSegment;
+        }
+    }
+
+    public static class ListExtension
+    {
+        public static float Percentile(this IEnumerable<float> source, float percentile)
+        {
+            if (percentile < 0 || percentile > 1)
+                throw new ArgumentOutOfRangeException(nameof(percentile), "Percentile must be between 0 and 1");
+
+            var sorted = source.OrderBy(n => n).ToList();
+            int count = sorted.Count;
+            if (count == 0)
+                throw new InvalidOperationException("Empty collection");
+
+            float position = (percentile) * (count - 1);
+            int lowerIndex = (int)Math.Floor(position);
+            int upperIndex = (int)Math.Ceiling(position);
+            float weight = position - lowerIndex;
+
+            if (upperIndex >= count)
+                return sorted[lowerIndex];
+
+            return sorted[lowerIndex] * (1 - weight) + sorted[upperIndex] * weight;
         }
     }
 }
